Only show concepts in warning if KG FF on (#659)
* Update trunk.yaml

* Only show topics filter if KG FF on

* Squashed commit of the following:

commit 3783ad504f19c5c5d6f63e84c765b9f71e5b1325
Merge: 1e6a767 8ac5345
Author: Katy Baulch <46493669+katybaulch@users.noreply.github.com>
Date:   Thu Jul 17 17:07:50 2025 +0100

    Merge branch 'main' into update-search-tests

commit 1e6a767dbd07af1eb3399b4a322322c4d55ae468
Author: Katy Baulch <46493669+katybaulch@users.noreply.github.com>
Date:   Thu Jul 17 16:49:38 2025 +0100

    Add sort and search option unit tests

commit 8ac5345be7982d1cdc6dccf6317d1d00103290df
Author: James Gorrie <jamesgorrie@users.noreply.github.com>
Date:   Thu Jul 17 14:59:37 2025 +0100

    feat: add principal laws and jurisdictions to litigation filter slideouts (#660)

commit b906f2f90037444a9f56e3ee97bfe7ca6565a210
Author: Katy Baulch <46493669+katybaulch@users.noreply.github.com>
Date:   Thu Jul 17 14:53:23 2025 +0100

    Split up KG search tests from core search

commit f2e10281744bf5002ab3a73aca793f09e0433dbe
Author: Katy Baulch <46493669+katybaulch@users.noreply.github.com>
Date:   Thu Jul 17 12:21:29 2025 +0100

    Only show topics filter if KG FF on

* Create featureFlags.tsx

* Fix imports
