+++ .trunk/trunk.yaml
@@ -3,14 +3,14 @@
 # To learn more about the format of this file, see https://docs.trunk.io/reference/trunk-yaml
 version: 0.1
 cli:
-  version: 1.21.0
+  version: 1.24.0
 
 # Trunk provides extensibility via plugins.
 # (https://docs.trunk.io/plugins)
 plugins:
   sources:
     - id: trunk
-      ref: v1.4.5
+      ref: v1.7.1
       uri: https://github.com/trunk-io/plugins
 
 # Many linters and tools depend on runtimes - configure them here.
@@ -18,7 +18,7 @@ plugins:
 runtimes:
   enabled:
     - go@1.21.0
-    - node@18.12.1
+    - node@22.16.0
     - python@3.10.8
 
 # This is the section where you manage your linters.
@@ -37,20 +37,20 @@ lint:
         - LICENSE.md
 
   enabled:
-    - actionlint@1.7.5
-    - checkov@3.2.46
-    - eslint@8
+    - actionlint@1.7.7
+    - checkov@3.2.451
+    - eslint@8.57.0
     - git-diff-check
-    - markdownlint@0.39.0
-    - oxipng@9.0.0
-    - prettier@3.2.5
+    - markdownlint@0.45.0
+    - oxipng@9.1.5
+    - prettier@3.6.2
     - shellcheck@0.10.0
     - shfmt@3.6.0
-    - svgo@3.2.0
+    - svgo@4.0.0
     - terrascan@1.19.1
-    - trivy@0.50.0
-    - trufflehog@3.71.0
-    - yamllint@1.35.1
+    - trivy@0.64.1
+    - trufflehog@3.90.0
+    - yamllint@1.37.1
 actions:
   disabled:
     - trunk-announce
